Remove the use of GlobalKey (#8015)
We no longer support syncing tables without primary key, so this can be
removed. However the management of tombstones did somewhat rely on this,
so a more primitive approach where we just make a linear search for
potential tombstone that must be revived. This is  of course slow, but
since this in on the local_only branch, tombstones are only used in some
copy operation.